Resolve reading app props in non-spring managed instances, implement first version of refresh tokens

+++ src/main/java/com/raven/api/service/impl/UserServiceImpl.java
@@ -1,6 +1,13 @@
 package com.raven.api.service.impl;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.raven.api.exception.EntryNotFoundException;
+import com.raven.api.exception.MissingTokenException;
 import com.raven.api.exception.NotLoggedInException;
 import com.raven.api.model.Role;
 import com.raven.api.model.User;
@@ -11,12 +18,24 @@ import com.raven.api.security.jwt.AuthUtils;
 import com.raven.api.service.UserService;
 import lombok.RequiredArgsConstructor;
 
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.support.MessageSourceAccessor;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.http.MediaType;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @Service
 @RequiredArgsConstructor
@@ -29,6 +48,17 @@ public class UserServiceImpl implements UserService {
     private final MessageSourceAccessor accessor;
 
     private final PasswordEncoder passwordEncoder;
+
+    private final String TOKEN_TYPE = "Bearer ";
+
+    @Value(value = "${jwt.secret}")
+	private String secret;
+
+    @Value(value = "${jwt.claim}")
+    private String claim;
+
+    @Value(value = "${jwt.access-token-expiration-time-millis}")
+	private Long accessTokenExpirationTimeMillis;
 
     @Override
     public User createUser(User user, RoleName roleName) {
@@ -90,7 +120,7 @@ public class UserServiceImpl implements UserService {
         final Optional<String> username = AuthUtils.getCurrentUserUsername();
         
         if (username.isEmpty()) {
-            throw new NotLoggedInException(accessor.getMessage("noLogin"));
+            throw new NotLoggedInException(accessor.getMessage("user.noLogin"));
         }
 
         return findUserByUsername(username.get());
@@ -98,8 +128,49 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long id) {
+        if (id == null) {
+            throw new EntryNotFoundException(accessor.getMessage("user.id.empty"));
+        }
         this.userRepository.deleteById(id);
     }
 
+    @Override
+    public void refreshToken(HttpServletRequest request, HttpServletResponse response) {
+        String authorizationHeader = request.getHeader(AUTHORIZATION);
+
+            if (authorizationHeader != null && authorizationHeader.startsWith(this.TOKEN_TYPE)) {
+                try {
+                    String refreshToken = authorizationHeader.substring(TOKEN_TYPE.length());
+                    Algorithm algorithm = Algorithm.HMAC256(this.secret.getBytes());
+                    JWTVerifier verifier = JWT.require(algorithm).build();
+                    DecodedJWT decodedJWT = verifier.verify(refreshToken);
+                    String username = decodedJWT.getSubject();
+                    User user = this.findUserByUsername(username);
+                    String accessToken = JWT.create()
+                        .withSubject(user.getUsername())
+                        .withExpiresAt(new Date(this.accessTokenExpirationTimeMillis))
+                        .withIssuer(request.getRequestURL().toString())
+                        .withClaim(this.claim, user.getRoles().stream().map(Role::getRoleName).collect(Collectors.toList()))
+                        .sign(algorithm);
+
+
+                    Map<String, String> tokens = new HashMap<>();
+                    tokens.put("access_token", accessToken);
+                    tokens.put("refresh_token", refreshToken);
+                    response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+                    new ObjectMapper().writeValue(response.getOutputStream(), tokens);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } catch (JWTVerificationException e) {
+                    try {
+                        response.sendError(FORBIDDEN.value());
+                    } catch (IOException e1) {
+                        e1.printStackTrace();
+                    }
+                }
+            } else {
+                throw new MissingTokenException("Missing authorization header or refresh token.");
+            }        
+    }
 
 }
